Refactor: api type 분리
jindonyy/issue-tracker/#26
