+++ FE/src/api/type.ts
@@ -1,0 +1,12 @@
+import { AxiosRequestConfig } from 'axios';
+
+type IssueStateType = 'open' | 'close';
+
+type AxiosType = {
+  method: 'get' | 'post' | 'put' | 'delete' | 'patch';
+  url: string;
+  data?: { [key: string]: string };
+  config?: AxiosRequestConfig;
+};
+
+export type { IssueStateType, AxiosType };
